Remove utils from webap
